before to move write task to same thread of device

+++ ui/filessystemwidget.cpp
@@ -2,6 +2,7 @@
 #include "ui_filessystemwidget.h"
 #include <QFileDialog>
 #include <QPair>
+#include "../core/devicefilessystem.h"
 
 FilesSystemWidget::FilesSystemWidget(Device* device,QWidget *parent) :
     QWidget(parent),_device(device),
@@ -9,7 +10,10 @@ FilesSystemWidget::FilesSystemWidget(Device* device,QWidget *parent) :
 {
     ui->setupUi(this);
     QObject::connect(_device->GetFileSystem(),&DeviceFilesSystem::FileListUpdated,this,&FilesSystemWidget::OnFileListUpdated);
+    QObject::connect(_device->GetFileSystem(),&DeviceFilesSystem::UploadFileFailed,this,&FilesSystemWidget::OnFileListUpdated);
+    QObject::connect(_device->GetFileSystem(),&DeviceFilesSystem::WaitListUpdated,this,&FilesSystemWidget::OnFileListUpdated);
     ui->_delete_file_button->setVisible(false);
+    OnFileListUpdated();
 }
 
 FilesSystemWidget::~FilesSystemWidget()
@@ -29,20 +33,45 @@ void FilesSystemWidget::OnFileListUpdated()
     auto files=_device->GetFileSystem()->GetFileList();
     for(int i=0;i<files.size();i++){
         QListWidgetItem* lwi=new QListWidgetItem(ui->_files_list);
+        lwi->setIcon(QIcon(":/icon/images/file.png"));
         lwi->setText(QString(files.keys()[i]+" : "+QByteArray::number(files.values()[i])));
+        ui->_files_list->addItem(lwi);
+    }
+
+    auto wfiles=_device->GetFileSystem()->GetWaitUploadingList();
+    for(int i=0;i<wfiles.size();i++){
+        QListWidgetItem* lwi=new QListWidgetItem(ui->_files_list);
+        lwi->setIcon(QIcon(":/icon/images/uploading_file.png"));
+        lwi->setText(wfiles[i]+"*");
+        ui->_files_list->addItem(lwi);
+    }
+
+    auto ffiles=_device->GetFileSystem()->GetFailedUploads();
+    for(int i=0;i<ffiles.size();i++){
+        QListWidgetItem* lwi=new QListWidgetItem(ui->_files_list);
+        lwi->setIcon(QIcon(":/icon/images/fail.png"));
+        lwi->setText(ffiles[i]+"*");
         ui->_files_list->addItem(lwi);
     }
 }
 
 void FilesSystemWidget::on__files_list_itemSelectionChanged()
 {
-    qDebug()<<"item selected";
+    ui->_delete_file_button->setVisible(false);
+    ui->_stop_upload_button->setVisible(false);
     if(ui->_files_list->selectedItems().length()>0){
-        ui->_delete_file_button->setVisible(true);
-    }
-    else
-    {
-        ui->_delete_file_button->setVisible(false);
+        for(QListWidgetItem* &item:ui->_files_list->selectedItems()){
+            QString name=item->text().mid(0,item->text().indexOf(" "));
+            if(_device->GetFileSystem()->GetFileList().contains(name.toUtf8())){
+                ui->_delete_file_button->setVisible(true);
+            }
+            else if(name.indexOf("*")){
+                if(_device->GetFileSystem()->GetWaitUploadingList().contains(name.mid(0,name.length()-1).toUtf8()))
+                {
+                    ui->_stop_upload_button->setVisible(true);
+                }
+            }
+        }
     }
 }
 
@@ -51,7 +80,9 @@ void FilesSystemWidget::on__delete_file_button_clicked()
     ui->_delete_file_button->setVisible(false);
     for(QListWidgetItem* &item:ui->_files_list->selectedItems()){
         QString name=item->text().mid(0,item->text().indexOf(" "));
-        _device->GetFileSystem()->DeleteFile(name.toUtf8());
+        if(_device->GetFileSystem()->GetFileList().contains(name.toUtf8())){
+            _device->GetFileSystem()->DeleteFile(name.toUtf8());
+        }
     }
 
 }
@@ -59,16 +90,36 @@ void FilesSystemWidget::on__delete_file_button_clicked()
 void FilesSystemWidget::on__upload_file_button_clicked()
 {
     QFileDialog fd;
-    fd.setFileMode(QFileDialog::FileMode::ExistingFile);
+    fd.setFileMode(QFileDialog::FileMode::ExistingFiles);
     fd.setNameFilter("*.GCO *.gcode");
     if(fd.exec()){
         if(fd.selectedFiles().length()>0)
         {
-            QFile file(fd.selectedFiles()[0]);
-            qDebug()<<QUrl(file.fileName()).fileName();
-            file.open(QIODevice::ReadOnly);
-            QByteArray ba=file.readAll();
-            _device->GetFileSystem()->UploadFile(QUrl(file.fileName()).fileName().replace("gcode","GCO").toUpper().toUtf8(),ba);
+            for(QString f:fd.selectedFiles())
+            {
+                QFile file(f);
+                QByteArray fileName=QUrl(file.fileName()).fileName().replace("gcode","GCO").toUpper().toUtf8();
+                if(!ui->_overwrite_checkbox->isChecked() && _device->GetFileSystem()->GetFileList().contains(fileName))
+                    continue;
+                qDebug()<<QUrl(file.fileName()).fileName();
+                file.open(QIODevice::ReadOnly);
+                QByteArray ba=file.readAll();
+                _device->GetFileSystem()->UploadFile(fileName,ba);
+            }
         }
     }
+}
+
+
+void FilesSystemWidget::on__stop_upload_button_clicked()
+{
+    for(QListWidgetItem* item:ui->_files_list->selectedItems()){
+        QString name=item->text().mid(0,item->text().indexOf(" "));
+        if(_device->GetFileSystem()->GetWaitUploadingList().contains(name.toUtf8().mid(0,name.length()-1))){
+            _device->GetFileSystem()->StopUpload(name.toUtf8().mid(0,name.length()-1));
+            qDebug()<<"stop upload called";
+        }
+    }
+
+    OnFileListUpdated();
 }
